gaia/miner/inference_service/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
PyYAML

# Add your project-specific dependencies here, for example:
# xarray
# pandas
# numpy
# scikit-learn
# torch
# torchvision
# torchaudio
# numcodecs
# zarr
# base64
# pickle

# Potentially aurora-weather (or the correct package name for the Aurora SDK)
# If Aurora requires specific system libraries, they need to be in the Dockerfile 